+++ backend/internal/auth/jwt.go
@@ -7,7 +7,7 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 )
 
-func GenerateJWT(userID int64) (string, error) {
+func GenerateJWT(userID uint) (string, error) {
 	jwtSecret := []byte(os.Getenv("JWT_SECRET"))
 	claims := jwt.MapClaims{
 		"user_id": userID,
@@ -18,7 +18,7 @@ func GenerateJWT(userID int64) (string, error) {
 	return token.SignedString(jwtSecret)
 }
 
-func GenerateRefreshToken(userID int64) (string, error) {
+func GenerateRefreshToken(userID uint) (string, error) {
 	jwtSecret := []byte(os.Getenv("JWT_REFRESH_SECRET"))
 	claims := jwt.MapClaims{
 		"user_id": userID,
